requirements.txt
pandas
scikit-learn
tqdm
google-generativeai
python-dotenv
transformers
huggingface-hub
torch
psutil
imblearn
tenacity
fastapi
uvicorn
kaggle
click
